+++ CHANGELOG.md
@@ -25,6 +25,12 @@ You should also add project tags for each release in Github, see [Managing relea
 - Modify `doc_retriever` script to store in Label Studio formatted Json.
 - Refactor parsing arguments in both `doc_retriever` and `label_studio_uploader`
 - Modify `doc_retriever` script to accept optional collection ID
+- Refactor multilabel handling in `preprocessing` 
+- Adding multilabel classification support in `bert_recipe`
+- Adding multilabel classification support in `sk_naive_bayes_recipe`
+- Modify `trainer` to use metadata from preprocessing more effectively
+- Modify `trainer` to handle multilabel classification
+- Modify `evaluation` to handle multilabel classification
 - Modify `evaluation` to accept optional args in main